fix(serve): keep replay cap validation out of fast path runtime
